+++ blog.wwawing.com/gatsby-node.js
@@ -1,54 +1,16 @@
-const path = require(`path`)
 const { createFilePath } = require(`gatsby-source-filesystem`)
 
-exports.createPages = async ({ graphql, actions }) => {
-  const { createPage } = actions
+require(`ts-node`).register({
+  compilerOptions: {
+    module: 'commonjs',
+    target: 'es2017',
+    noImplicitAny: false,
+  },
+})
 
-  const blogPost = path.resolve(`./src/templates/blog-post.tsx`)
-  const result = await graphql(
-    `
-      query BlogPosts {
-        allMarkdownRemark(
-          sort: { fields: [frontmatter___date], order: DESC }
-          limit: 1000
-        ) {
-          edges {
-            node {
-              fields {
-                slug
-              }
-              frontmatter {
-                title
-              }
-            }
-          }
-        }
-      }
-    `
-  )
-
-  if (result.errors) {
-    throw result.errors
-  }
-
-  // Create blog posts pages.
-  const posts = result.data.allMarkdownRemark.edges
-
-  posts.forEach((post, index) => {
-    const previous = index === posts.length - 1 ? null : posts[index + 1].node
-    const next = index === 0 ? null : posts[index - 1].node
-
-    createPage({
-      path: post.node.fields.slug,
-      component: blogPost,
-      context: {
-        slug: post.node.fields.slug,
-        previous,
-        next,
-      },
-    })
-  })
-}
+// FIXME: We've encountered an error: Objects are not valid as a React child (found: GraphQLDocumentError: GraphQLDocumentError: Unknown fragment "GatsbyImageSharpFixed".).
+//    If you meant to render a collection of children, use an array instead.
+exports.createPages = require(`./gatsby-node/create-pages`).createPages
 
 exports.onCreateNode = ({ node, actions, getNode }) => {
   const { createNodeField } = actions
