+++ blog.wwawing.com/package.json
@@ -40,7 +40,8 @@
     "typography-theme-bootstrap": "^0.16.19"
   },
   "devDependencies": {
-    "prettier": "2.0.5"
+    "prettier": "2.0.5",
+    "ts-node": "^9.0.0"
   },
   "homepage": "https://github.com/gatsbyjs/gatsby-starter-blog#readme",
   "keywords": [
